move codecs out again
