+++ src/aioway/io/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
-from .codecs import *
 from .dsets import *
 from .envs import *
 from .sess import *
