+++ tests/codecs/test_collate.py
@@ -5,7 +5,7 @@ import typing
 import pytest
 import torch
 
-from aioway.io import chunk_collate_decompose_dim
+from aioway.codecs import chunk_collate_decompose_dim
 
 
 class SamplesCollateDim(typing.NamedTuple):
